- add client layer - rename WeatherService to WeatherSearch - use int as temperature in place of double

+++ weatherService/src/main/java/com/nprogramming/weather/utils/TemperatureHelper.java
@@ -8,6 +8,6 @@ public final class TemperatureHelper {
 
         return new BigDecimal(value)
                 .setScale(2, BigDecimal.ROUND_HALF_EVEN)
-                .doubleValue();
+                .intValue();
     }
 }
